ingestion_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
cassandra-driver==3.29.1      # Datastax driver; works with Cassandra 4.x and ScyllaDB
asyncpg==0.29.0               # Async Postgres driver; faster than psycopg2 for FastAPI
pydantic==2.7.1               # Data validation; FastAPI uses this internally
python-dotenv==1.0.1          # Load .env in local runs outside Docker
prometheus-client==0.20.0     # /metrics endpoint for observability